initial commit.
calvad_outliers is the main calling point of this code.

get_detectors just wraps the couch_cacher code, and get_time is a bit
of code that converts year into start and end times for a couchdb
query.

No tests yet, code is not close to operational, but getting there.